@@ -1,36 +1,33 @@
-/**
- * Copyright (c) 2019-present, GraphQL Foundation
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow strict
- */
+// Copyright (c) 2019-present, GraphQL Foundation
+//
+// This source code is licensed under the MIT license found in the
+// LICENSE file in the root directory of this source tree.
 
 // A Function, which when given an Array of keys, returns a Promise of an Array
 // of values or Errors.
+
 export type BatchLoadFn<K, V> = (
-  keys: $ReadOnlyArray<K>,
-) => Promise<$ReadOnlyArray<V | Error>>;
+  keys: ReadonlyArray<K>,
+) => Promise<ReadonlyArray<V | Error>>;
 
 // Optionally turn off batching or caching or provide a cache key function or a
 // custom cache instance.
-export type Options<K, V, C = K> = {
-  batch?: boolean,
-  maxBatchSize?: number,
-  batchScheduleFn?: (callback: () => void) => void,
-  cache?: boolean,
-  cacheKeyFn?: (key: K) => C,
-  cacheMap?: CacheMap<C, Promise<V>> | null,
-  name?: string,
+export type Options<K, V, C> = {
+  batch?: boolean;
+  maxBatchSize?: number;
+  batchScheduleFn?: (callback: () => void) => void;
+  cache?: boolean;
+  cacheKeyFn?: (key: K) => C;
+  cacheMap?: CacheMap<C, Promise<V>> | null;
+  name?: string;
 };
 
 // If a custom cache is provided, it must be of this type (a subset of ES6 Map).
 export type CacheMap<K, V> = {
-  get(key: K): V | void,
-  set(key: K, value: V): any,
-  delete(key: K): any,
-  clear(): any,
+  get(key: K): V | undefined;
+  set(key: K, value: V): void;
+  delete(key: K): void;
+  clear(): void;
 };
 
 /**
@@ -44,6 +41,14 @@ export type CacheMap<K, V> = {
  * web request.
  */
 class DataLoader<K, V, C = K> {
+  private _batchLoadFn: BatchLoadFn<K, V>;
+  private _maxBatchSize: number;
+  private _batchScheduleFn: (callback: () => void) => void;
+  private _cacheKeyFn: (key: K) => C;
+  private _cacheMap: CacheMap<C, Promise<V>> | null;
+  private _batch: Batch<K, V> | null;
+  public name: string | null;
+
   constructor(batchLoadFn: BatchLoadFn<K, V>, options?: Options<K, V, C>) {
     if (typeof batchLoadFn !== 'function') {
       throw new TypeError(
@@ -60,14 +65,126 @@ class DataLoader<K, V, C = K> {
     this.name = getValidName(options);
   }
 
-  // Private
-  _batchLoadFn: BatchLoadFn<K, V>;
-  _maxBatchSize: number;
-  _batchScheduleFn: (() => void) => void;
-  _cacheKeyFn: K => C;
-  _cacheMap: CacheMap<C, Promise<V>> | null;
-  _batch: Batch<K, V> | null;
+  private dispatchBatch(batch: Batch<K, V>) {
+    // Mark this batch as having been dispatched.
+    batch.hasDispatched = true;
 
+    // If there's nothing to load, resolve any cache hits and return early.
+    if (batch.keys.length === 0) {
+      resolveCacheHits(batch);
+      return;
+    }
+
+    // Call the provided batchLoadFn for this loader with the batch's keys and
+    // with the loader as the `this` context.
+    let batchPromise: Promise<ReadonlyArray<V | Error>>;
+    try {
+      batchPromise = this._batchLoadFn(batch.keys);
+    } catch (e) {
+      return this.failedDispatch(
+        batch,
+        new TypeError(
+          'DataLoader must be constructed with a function which accepts ' +
+            'Array<key> and returns Promise<Array<value>>, but the function ' +
+            `errored synchronously: ${String(e)}.`,
+        ),
+      );
+    }
+
+    // Assert the expected response from batchLoadFn
+    if (!batchPromise || typeof batchPromise.then !== 'function') {
+      return this.failedDispatch(
+        batch,
+        new TypeError(
+          'DataLoader must be constructed with a function which accepts ' +
+            'Array<key> and returns Promise<Array<value>>, but the function did ' +
+            `not return a Promise: ${String(batchPromise)}.`,
+        ),
+      );
+    }
+
+    // Await the resolution of the call to batchLoadFn.
+    batchPromise
+      .then(values => {
+        // Assert the expected resolution from batchLoadFn.
+        if (!isArrayLike(values)) {
+          throw new TypeError(
+            'DataLoader must be constructed with a function which accepts ' +
+              'Array<key> and returns Promise<Array<value>>, but the function did ' +
+              `not return a Promise of an Array: ${String(values)}.`,
+          );
+        }
+        if (values.length !== batch.keys.length) {
+          throw new TypeError(
+            'DataLoader must be constructed with a function which accepts ' +
+              'Array<key> and returns Promise<Array<value>>, but the function did ' +
+              'not return a Promise of an Array of the same length as the Array ' +
+              'of keys.' +
+              `\n\nKeys:\n${String(batch.keys)}` +
+              `\n\nValues:\n${String(values)}`,
+          );
+        }
+
+        // Resolve all cache hits in the same micro-task as freshly loaded values.
+        resolveCacheHits(batch);
+
+        // Step through values, resolving or rejecting each Promise in the batch.
+        for (let i = 0; i < batch.callbacks.length; i++) {
+          const value = values[i]!;
+          if (value instanceof Error) {
+            batch.callbacks[i]!.reject(value);
+          } else {
+            batch.callbacks[i]!.resolve(value);
+          }
+        }
+      })
+      .catch(error => {
+        this.failedDispatch(batch, error);
+      });
+  }
+
+  // Private: Either returns the current batch, or creates and schedules a
+  // dispatch of a new batch for the given loader.
+  private getCurrentBatch(): Batch<K, V> {
+    // If there is an existing batch which has not yet dispatched and is within
+    // the limit of the batch size, then return it.
+    const existingBatch = this._batch;
+    if (
+      existingBatch !== null &&
+      !existingBatch.hasDispatched &&
+      existingBatch.keys.length < this._maxBatchSize
+    ) {
+      return existingBatch;
+    }
+
+    // Otherwise, create a new batch for this loader.
+    const newBatch: Batch<K, V> = {
+      hasDispatched: false,
+      keys: [],
+      callbacks: [],
+    };
+
+    // Store it on the loader so it may be reused.
+    this._batch = newBatch;
+
+    // Then schedule a task to dispatch this batch of requests.
+    this._batchScheduleFn(() => {
+      this.dispatchBatch(newBatch);
+    });
+
+    return newBatch;
+  }
+
+  // Private: do not cache individual loads if the entire batch dispatch fails,
+  // but still reject each request so they do not hang.
+  private failedDispatch(batch: Batch<K, V>, error: Error) {
+    // Cache hits are resolved, even though the batch failed.
+    resolveCacheHits(batch);
+    for (let i = 0; i < batch.keys.length; i++) {
+      this.clear(batch.keys[i]!);
+      batch.callbacks[i]!.reject(error);
+    }
+  }
   /**
    * Loads a key, returning a `Promise` for the value represented by that key.
    */
@@ -79,9 +196,9 @@ class DataLoader<K, V, C = K> {
       );
     }
 
-    const batch = getCurrentBatch(this);
+    const batch = this.getCurrentBatch();
     const cacheMap = this._cacheMap;
-    let cacheKey: ?C;
+    let cacheKey: C | undefined;
 
     // If caching and there is a cache-hit, return cached Promise.
     if (cacheMap) {
@@ -100,13 +217,13 @@ class DataLoader<K, V, C = K> {
     // Otherwise, produce a new Promise for this key, and enqueue it to be
     // dispatched along with the current batch.
     batch.keys.push(key);
-    const promise = new Promise((resolve, reject) => {
+    const promise = new Promise<V>((resolve, reject) => {
       batch.callbacks.push({ resolve, reject });
     });
 
     // If caching, cache this promise.
-    if (cacheMap) {
-      cacheMap.set((cacheKey: any), promise);
+    if (cacheMap && cacheKey) {
+      cacheMap.set(cacheKey, promise);
     }
 
     return promise;
@@ -132,17 +249,17 @@ class DataLoader<K, V, C = K> {
    *     // c instanceof Error
    *
    */
-  loadMany(keys: $ReadOnlyArray<K>): Promise<Array<V | Error>> {
+  loadMany(keys: ReadonlyArray<K>): Promise<Array<V | Error>> {
     if (!isArrayLike(keys)) {
       throw new TypeError(
         'The loader.loadMany() function must be called with Array<key> ' +
-          `but got: ${(keys: any)}.`,
+          `but got: ${keys}.`,
       );
     }
     // Support ArrayLike by using only minimal property access
-    const loadPromises = [];
+    const loadPromises: Array<Promise<V | Error>> = [];
     for (let i = 0; i < keys.length; i++) {
-      loadPromises.push(this.load(keys[i]).catch(error => error));
+      loadPromises.push(this.load(keys[i]!).catch(error => error));
     }
     return Promise.all(loadPromises);
   }
@@ -188,7 +305,7 @@ class DataLoader<K, V, C = K> {
       if (cacheMap.get(cacheKey) === undefined) {
         // Cache a rejected promise if the value is an Error, in order to match
         // the behavior of load(key).
-        let promise;
+        let promise: Promise<V>;
         if (value instanceof Error) {
           promise = Promise.reject(value);
           // Since this is a case where an Error is intentionally being primed
@@ -200,15 +317,9 @@ class DataLoader<K, V, C = K> {
         cacheMap.set(cacheKey, promise);
       }
     }
+
     return this;
   }
-
-  /**
-   * The name given to this `DataLoader` instance. Useful for APM tools.
-   *
-   * Is `null` if not set in the constructor.
-   */
-  name: string | null;
 }
 
 // Private: Enqueue a Job to be executed after all "PromiseJobs" Jobs.
@@ -237,8 +348,8 @@ class DataLoader<K, V, C = K> {
 // for enqueuing a job to be performed after promise microtasks and before the
 // next macrotask. For browser environments, a macrotask is used (via
 // setImmediate or setTimeout) at a potential performance penalty.
-const enqueuePostPromiseJob =
-  typeof process === 'object' && typeof process.nextTick === 'function'
+function getEnqueuePostPromiseJob(): (fn: () => void) => void {
+  return typeof process === 'object' && typeof process.nextTick === 'function'
     ? function (fn) {
         if (!resolvedPromise) {
           resolvedPromise = Promise.resolve();
@@ -248,164 +359,39 @@ const enqueuePostPromiseJob =
         });
       }
     : typeof setImmediate === 'function'
-    ? function (fn) {
-        setImmediate(fn);
-      }
-    : function (fn) {
-        setTimeout(fn);
-      };
+      ? function (fn) {
+          setImmediate(fn);
+        }
+      : function (fn) {
+          setTimeout(fn);
+        };
+}
 
 // Private: cached resolved Promise instance
-let resolvedPromise;
+let resolvedPromise: Promise<void> | undefined;
 
 // Private: Describes a batch of requests
 type Batch<K, V> = {
-  hasDispatched: boolean,
-  keys: Array<K>,
+  hasDispatched: boolean;
+  keys: Array<K>;
   callbacks: Array<{
-    resolve: (value: V) => void,
-    reject: (error: Error) => void,
-  }>,
-  cacheHits?: Array<() => void>,
+    resolve: (value: V) => void;
+    reject: (error: Error) => void;
+  }>;
+  cacheHits?: Array<() => void>;
 };
 
-// Private: Either returns the current batch, or creates and schedules a
-// dispatch of a new batch for the given loader.
-function getCurrentBatch<K, V>(loader: DataLoader<K, V, any>): Batch<K, V> {
-  // If there is an existing batch which has not yet dispatched and is within
-  // the limit of the batch size, then return it.
-  const existingBatch = loader._batch;
-  if (
-    existingBatch !== null &&
-    !existingBatch.hasDispatched &&
-    existingBatch.keys.length < loader._maxBatchSize
-  ) {
-    return existingBatch;
-  }
-
-  // Otherwise, create a new batch for this loader.
-  const newBatch = { hasDispatched: false, keys: [], callbacks: [] };
-
-  // Store it on the loader so it may be reused.
-  loader._batch = newBatch;
-
-  // Then schedule a task to dispatch this batch of requests.
-  loader._batchScheduleFn(() => {
-    dispatchBatch(loader, newBatch);
-  });
-
-  return newBatch;
-}
-
-function dispatchBatch<K, V>(
-  loader: DataLoader<K, V, any>,
-  batch: Batch<K, V>,
-) {
-  // Mark this batch as having been dispatched.
-  batch.hasDispatched = true;
-
-  // If there's nothing to load, resolve any cache hits and return early.
-  if (batch.keys.length === 0) {
-    resolveCacheHits(batch);
-    return;
-  }
-
-  // Call the provided batchLoadFn for this loader with the batch's keys and
-  // with the loader as the `this` context.
-  let batchPromise;
-  try {
-    batchPromise = loader._batchLoadFn(batch.keys);
-  } catch (e) {
-    return failedDispatch(
-      loader,
-      batch,
-      new TypeError(
-        'DataLoader must be constructed with a function which accepts ' +
-          'Array<key> and returns Promise<Array<value>>, but the function ' +
-          `errored synchronously: ${String(e)}.`,
-      ),
-    );
-  }
-
-  // Assert the expected response from batchLoadFn
-  if (!batchPromise || typeof batchPromise.then !== 'function') {
-    return failedDispatch(
-      loader,
-      batch,
-      new TypeError(
-        'DataLoader must be constructed with a function which accepts ' +
-          'Array<key> and returns Promise<Array<value>>, but the function did ' +
-          `not return a Promise: ${String(batchPromise)}.`,
-      ),
-    );
-  }
-
-  // Await the resolution of the call to batchLoadFn.
-  batchPromise
-    .then(values => {
-      // Assert the expected resolution from batchLoadFn.
-      if (!isArrayLike(values)) {
-        throw new TypeError(
-          'DataLoader must be constructed with a function which accepts ' +
-            'Array<key> and returns Promise<Array<value>>, but the function did ' +
-            `not return a Promise of an Array: ${String(values)}.`,
-        );
-      }
-      if (values.length !== batch.keys.length) {
-        throw new TypeError(
-          'DataLoader must be constructed with a function which accepts ' +
-            'Array<key> and returns Promise<Array<value>>, but the function did ' +
-            'not return a Promise of an Array of the same length as the Array ' +
-            'of keys.' +
-            `\n\nKeys:\n${String(batch.keys)}` +
-            `\n\nValues:\n${String(values)}`,
-        );
-      }
-
-      // Resolve all cache hits in the same micro-task as freshly loaded values.
-      resolveCacheHits(batch);
-
-      // Step through values, resolving or rejecting each Promise in the batch.
-      for (let i = 0; i < batch.callbacks.length; i++) {
-        const value = values[i];
-        if (value instanceof Error) {
-          batch.callbacks[i].reject(value);
-        } else {
-          batch.callbacks[i].resolve(value);
-        }
-      }
-    })
-    .catch(error => {
-      failedDispatch(loader, batch, error);
-    });
-}
-
-// Private: do not cache individual loads if the entire batch dispatch fails,
-// but still reject each request so they do not hang.
-function failedDispatch<K, V>(
-  loader: DataLoader<K, V, any>,
-  batch: Batch<K, V>,
-  error: Error,
-) {
-  // Cache hits are resolved, even though the batch failed.
-  resolveCacheHits(batch);
-  for (let i = 0; i < batch.keys.length; i++) {
-    loader.clear(batch.keys[i]);
-    batch.callbacks[i].reject(error);
-  }
-}
-
 // Private: Resolves the Promises for any cache hits in this batch.
-function resolveCacheHits(batch: Batch<any, any>) {
+function resolveCacheHits<K, V>(batch: Batch<K, V>) {
   if (batch.cacheHits) {
     for (let i = 0; i < batch.cacheHits.length; i++) {
-      batch.cacheHits[i]();
+      batch.cacheHits[i]!();
     }
   }
 }
 
 // Private: given the DataLoader's options, produce a valid max batch size.
-function getValidMaxBatchSize(options: ?Options<any, any, any>): number {
+function getValidMaxBatchSize<K, V, C>(options?: Options<K, V, C>): number {
   const shouldBatch = !options || options.batch !== false;
   if (!shouldBatch) {
     return 1;
@@ -416,43 +402,48 @@ function getValidMaxBatchSize(options: ?Options<any, any, any>): number {
   }
   if (typeof maxBatchSize !== 'number' || maxBatchSize < 1) {
     throw new TypeError(
-      `maxBatchSize must be a positive number: ${(maxBatchSize: any)}`,
+      `maxBatchSize must be a positive number: ${maxBatchSize}`,
     );
   }
   return maxBatchSize;
 }
 
 // Private
-function getValidBatchScheduleFn(
-  options: ?Options<any, any, any>,
-): (() => void) => void {
+function getValidBatchScheduleFn<K, V, C>(
+  options?: Options<K, V, C>,
+): (callback: () => void) => void {
   const batchScheduleFn = options && options.batchScheduleFn;
   if (batchScheduleFn === undefined) {
-    return enqueuePostPromiseJob;
+    return getEnqueuePostPromiseJob();
   }
   if (typeof batchScheduleFn !== 'function') {
     throw new TypeError(
-      `batchScheduleFn must be a function: ${(batchScheduleFn: any)}`,
+      `batchScheduleFn must be a function: ${batchScheduleFn}`,
     );
   }
   return batchScheduleFn;
 }
 
 // Private: given the DataLoader's options, produce a cache key function.
-function getValidCacheKeyFn<K, C>(options: ?Options<K, any, C>): K => C {
+function getValidCacheKeyFn<K, V, C>(
+  options?: Options<K, V, C>,
+): (key: K) => C {
   const cacheKeyFn = options && options.cacheKeyFn;
   if (cacheKeyFn === undefined) {
-    return (key => key: any);
+    //
+    // TODO: Find a better way to express this cast.
+    //
+    return (key: K) => key as unknown as C;
   }
   if (typeof cacheKeyFn !== 'function') {
-    throw new TypeError(`cacheKeyFn must be a function: ${(cacheKeyFn: any)}`);
+    throw new TypeError(`cacheKeyFn must be a function: ${cacheKeyFn}`);
   }
   return cacheKeyFn;
 }
 
 // Private: given the DataLoader's options, produce a CacheMap to be used.
 function getValidCacheMap<K, V, C>(
-  options: ?Options<K, V, C>,
+  options?: Options<K, V, C>,
 ): CacheMap<C, Promise<V>> | null {
   const shouldCache = !options || options.cache !== false;
   if (!shouldCache) {
@@ -463,7 +454,7 @@ function getValidCacheMap<K, V, C>(
     return new Map();
   }
   if (cacheMap !== null) {
-    const cacheFunctions = ['get', 'set', 'delete', 'clear'];
+    const cacheFunctions = ['get', 'set', 'delete', 'clear'] as const;
     const missingFunctions = cacheFunctions.filter(
       fnName => cacheMap && typeof cacheMap[fnName] !== 'function',
     );
@@ -476,7 +467,7 @@ function getValidCacheMap<K, V, C>(
   return cacheMap;
 }
 
-function getValidName<K, V, C>(options: ?Options<K, V, C>): string | null {
+function getValidName<K, V, C>(options?: Options<K, V, C>): string | null {
   if (options && options.name) {
     return options.name;
   }
@@ -485,14 +476,15 @@ function getValidName<K, V, C>(options: ?Options<K, V, C>): string | null {
 }
 
 // Private
-function isArrayLike(x: mixed): boolean {
+function isArrayLike(x: unknown): x is { length: number } {
   return (
     typeof x === 'object' &&
     x !== null &&
+    'length' in x &&
     typeof x.length === 'number' &&
     (x.length === 0 ||
       (x.length > 0 && Object.prototype.hasOwnProperty.call(x, x.length - 1)))
   );
 }
 
-module.exports = DataLoader;
+export default DataLoader;
